Using the device (disk) key instead of diskObjectId Issue/38 (#41)
The "diskObjectId" is not static and changes when a VM is migrated to another ESXi server within a cluster.
Then the incremental copy will not work.
The device key is unique for a VM. In combination with the VM name, it is suitable as an identifier to map disks from VMware to Openstack volumes.
This PR changes the behavior of migratekit  to use the device key while maintaining a fallback if an incremental copy was done on a VM before these changes. It fixes #38